Refactor file uploads to use direct upload utility
Extract file upload logic into reusable `uploadFileDirect` helper,
removing duplication across copilot and workspace upload components.
This also eliminates unnecessary dependencies on WebSocket tokens and
environment configuration from the copilot page.
