+++ autogpt_platform/frontend/src/app/(platform)/copilot/useCopilotPage.ts
@@ -5,10 +5,9 @@ import {
   type getV2ListSessionsResponse,
 } from "@/app/api/__generated__/endpoints/chat/chat";
 import { toast } from "@/components/molecules/Toast/use-toast";
+import { uploadFileDirect } from "@/lib/direct-upload";
 import { useBreakpoint } from "@/lib/hooks/useBreakpoint";
-import { getWebSocketToken } from "@/lib/supabase/actions";
 import { useSupabase } from "@/lib/supabase/hooks/useSupabase";
-import { environment } from "@/services/environment";
 import { useQueryClient } from "@tanstack/react-query";
 import type { FileUIPart } from "ai";
 import { useEffect, useRef, useState } from "react";
@@ -129,49 +128,25 @@ export function useCopilotPage() {
     files: File[],
     sid: string,
   ): Promise<UploadedFile[]> {
-    // Upload directly to the Python backend, bypassing the Next.js serverless
-    // proxy.  Vercel's 4.5 MB function payload limit would reject larger files
-    // when routed through /api/workspace/files/upload.
-    const { token, error: tokenError } = await getWebSocketToken();
-    if (tokenError || !token) {
-      toast({
-        title: "Authentication error",
-        description: "Please sign in again.",
-        variant: "destructive",
-      });
-      return [];
-    }
-
-    const backendBase = environment.getAGPTServerBaseUrl();
-
     const results = await Promise.allSettled(
       files.map(async (file) => {
-        const formData = new FormData();
-        formData.append("file", file);
-        const url = new URL("/api/workspace/files/upload", backendBase);
-        url.searchParams.set("session_id", sid);
-        const res = await fetch(url.toString(), {
-          method: "POST",
-          headers: { Authorization: `Bearer ${token}` },
-          body: formData,
-        });
-        if (!res.ok) {
-          const err = await res.text();
+        try {
+          const data = await uploadFileDirect(file, sid);
+          if (!data.file_id) throw new Error("No file_id returned");
+          return {
+            file_id: data.file_id,
+            name: data.name || file.name,
+            mime_type: data.mime_type || "application/octet-stream",
+          } as UploadedFile;
+        } catch (err) {
           console.error("File upload failed:", err);
           toast({
             title: "File upload failed",
             description: file.name,
             variant: "destructive",
           });
-          throw new Error(err);
+          throw err;
         }
-        const data = await res.json();
-        if (!data.file_id) throw new Error("No file_id returned");
-        return {
-          file_id: data.file_id,
-          name: data.name || file.name,
-          mime_type: data.mime_type || "application/octet-stream",
-        } as UploadedFile;
       }),
     );
     return results
