+++ autogpt_platform/frontend/src/components/renderers/InputRenderer/base/standard/widgets/FileInput/useWorkspaceUpload.ts
@@ -1,15 +1,10 @@
-import {
-  usePostWorkspaceUploadFileToWorkspace,
-  useDeleteWorkspaceDeleteAWorkspaceFile,
-} from "@/app/api/__generated__/endpoints/workspace/workspace";
+import { useDeleteWorkspaceDeleteAWorkspaceFile } from "@/app/api/__generated__/endpoints/workspace/workspace";
 import { useToast } from "@/components/molecules/Toast/use-toast";
+import { uploadFileDirect } from "@/lib/direct-upload";
 import { parseWorkspaceFileID, buildWorkspaceURI } from "@/lib/workspace-uri";
 
 export function useWorkspaceUpload() {
   const { toast } = useToast();
-
-  const { mutateAsync: uploadMutation } =
-    usePostWorkspaceUploadFileToWorkspace();
 
   const { mutate: deleteMutation } = useDeleteWorkspaceDeleteAWorkspaceFile({
     mutation: {
@@ -24,11 +19,7 @@ export function useWorkspaceUpload() {
   });
 
   async function handleUploadFile(file: File) {
-    const response = await uploadMutation({ data: { file } });
-    if (response.status !== 200) {
-      throw new Error("Upload failed");
-    }
-    const d = response.data;
+    const d = await uploadFileDirect(file);
     return {
       file_name: d.name,
       size: d.size_bytes,
